rename modifiers more
